+++ plugins/alignments/src/LinearAlignmentsDisplay/components/useAlignmentsBase.ts
@@ -104,13 +104,13 @@ export interface LinearAlignmentsDisplayModel {
       {
         arcX1: Uint32Array
         arcX2: Uint32Array
-        arcColorTypes: Float32Array
+        arcColorTypes: Uint8Array
         arcShapeTypes: Uint8Array
         arcYBp: Uint32Array
         numArcs: number
         linePositions: Uint32Array
         lineYs: Float32Array
-        lineColorTypes: Float32Array
+        lineColorTypes: Uint8Array
         numLines: number
       }
     >
